fix: share timeline fade fitting between preview and export
Move fitFades into a browser-safe leaf while preserving both legacy exports. Timeline normalization, ffmpeg export, overlay preview and trim edits now use one implementation.

History shows server fitting introduced in fccd90a25 and preview alignment in b4b330c7c after shortened overlays ramped differently. The current bodies matched, but no guard prevented another independent copy. Pin both exports to the shared function and retain the existing boundary regression coverage.
