+++ server/services/videoTimeline/segments.js
@@ -32,6 +32,9 @@ import { existsSync } from 'fs';
 import { join, resolve as resolvePath, sep } from 'path';
 import { PATHS, isTopLevelEntryName } from '../../lib/fileUtils.js';
 import { ServerError } from '../../lib/errorHandler.js';
+import { fitFades } from '../../lib/videoTimelineFades.js';
+
+export { fitFades } from '../../lib/videoTimelineFades.js';
 
 export const TIMELINE_SCHEMA_VERSION = 2;
 
@@ -50,22 +53,6 @@ export const AUDIO_ASSET_KINDS = ['audio', 'music'];
 export const MAX_SEGMENTS = 200;
 export const MAX_OVERLAYS = 50;
 export const MAX_AUDIO_TRACKS = 20;
-/**
- * Shrink a fade pair that no longer fits `duration`, scaling both
- * proportionally so the author's balance survives. Non-throwing twin of
- * `clampFades`: that one rejects an over-long pair at persist time, this one
- * repairs a pair that a LATER change made too long (a probe-shortened bed, a
- * legacy trim). ffmpeg's `fade` renders the whole segment black when its start
- * time goes negative, so an unfitted pair is not a cosmetic problem.
- */
-export function fitFades(fadeInSec, fadeOutSec, duration) {
-  const fin = Math.max(0, Number(fadeInSec) || 0);
-  const fout = Math.max(0, Number(fadeOutSec) || 0);
-  const span = fin + fout;
-  if (span <= duration || span === 0) return { fadeInSec: fin, fadeOutSec: fout };
-  const scale = Math.max(0, duration) / span;
-  return { fadeInSec: fin * scale, fadeOutSec: fout * scale };
-}
 
 export const MAX_FADE_SEC = 30;
 export const MAX_STILL_SEC = 600;
